moved exerciseInputFields to common folder
