+++ frontend/src/components/common/ExerciseInputFields.tsx
