+++ frontend/src/components/pages/add-program/AddProgram.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom'
 import { useAuth0 } from '@auth0/auth0-react'
 import { Button, Typography, Box, styled } from '@mui/material'
 import { ProgramNameField } from 'components/common'
-import { ExerciseInputFields } from './components'
+import { ExerciseInputFields } from '../../common'
 import { useSnackbar } from 'context/SnackbarContext'
 import { type ProgramFormInputs } from 'types/ProgramFormInputs'
 import BACKEND_URL from '../../../constants'
