--- frontend/src/components/pages/add-program/components/index.ts
@@ -1,3 +1,0 @@
-import { ExerciseInputFields } from './ExerciseInputFields'
-
-export { ExerciseInputFields }
