+++ frontend/src/demo/pages/demo-add-program/index.ts
@@ -1,0 +1,3 @@
+import { DemoAddProgram } from './DemoAddProgram'
+
+export { DemoAddProgram }
